create a toolkit package to separate the logic from the main.go
